+++ tests/Unit/SocketTest.php
@@ -49,7 +49,7 @@ final class SocketTest extends TestCase
 	{
 		$nameValuePairEncoder = new NameValuePairEncoder();
 		$packetEncoder        = new PacketEncoder();
-		$connection           = new UnixDomainSocket( 'unix:///var/run/php-uds.sock' );
+		$connection           = new UnixDomainSocket( '/var/run/php-uds.sock' );
 
 		return new Socket( $connection, $packetEncoder, $nameValuePairEncoder );
 	}
@@ -59,7 +59,7 @@ final class SocketTest extends TestCase
 		$socket  = $this->getSocket();
 		$data    = [ 'test-key' => 'unit' ];
 		$request = new PostRequest(
-			realpath( __DIR__ . '/../Integration/Workers/worker.php' ),
+			dirname( __DIR__ ) . '/Integration/Workers/worker.php',
 			http_build_query( $data )
 		);
 
@@ -80,7 +80,7 @@ final class SocketTest extends TestCase
 		$socket    = $this->getSocket();
 		$data      = [ 'test-key' => 'unit' ];
 		$request   = new PostRequest(
-			realpath( __DIR__ . '/../Integration/Workers/worker.php' ),
+			dirname( __DIR__ ) . '/Integration/Workers/worker.php',
 			http_build_query( $data )
 		);
 
@@ -100,7 +100,7 @@ final class SocketTest extends TestCase
 		$socket  = $this->getSocket();
 		$data    = [ 'test-key' => 'unit' ];
 		$request = new PostRequest(
-			realpath( __DIR__ . '/../Integration/Workers/worker.php' ),
+			dirname( __DIR__ ) . '/Integration/Workers/worker.php',
 			http_build_query( $data )
 		);
 		$request->addResponseCallbacks(
@@ -122,7 +122,7 @@ final class SocketTest extends TestCase
 		$socket  = $this->getSocket();
 		$data    = [ 'test-key' => 'unit' ];
 		$request = new PostRequest(
-			realpath( __DIR__ . '/../Integration/Workers/worker.php' ),
+			dirname( __DIR__ ) . '/Integration/Workers/worker.php',
 			http_build_query( $data )
 		);
 		$request->addFailureCallbacks(