+++ CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a CHANGELOG](http://keepachangelog.com).
 
+## [1.3.0] - 2017-06-15
+
+### Changed
+
+* Replaced methods `getHost()` and `getPort()` with `getSocketAddress()` in interface `hollodotme\FastCGI\Interfaces\ConfiguresSocketConnection` - [#9]
+* The transport protocol `unix://` must be omitted for the first parameter of `hollodotme\FastCGI\SocketConnections\UnixDomainSocket`  
+  Only the socket path must be passed. - [#9]
+* Relaced `fsockopen()` with `stream_socket_client()` for connecting to php-fpm. - [#9]
+
 ## [1.2.0] - 2017-05-27
 
 ### Added
@@ -88,6 +97,7 @@ Based on [Pierrick Charron](https://github.com/adoy)'s [PHP-FastCGI-Client](http
  * Getters/Setters for connect timeout, read/write timeout, keep alive, socket persistence from `Client` (now part of the socket connection)
  * Method `Client->getValues()`
 
+[1.3.0]: https://github.com/hollodotme/fast-cgi-client/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/hollodotme/fast-cgi-client/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/hollodotme/fast-cgi-client/compare/v1.0.1...v1.1.0
 [1.0.1]: https://github.com/hollodotme/fast-cgi-client/compare/v1.0.0...v1.0.1
@@ -97,3 +107,4 @@ Based on [Pierrick Charron](https://github.com/adoy)'s [PHP-FastCGI-Client](http
 [#2]: https://github.com/hollodotme/fast-cgi-client/issues/2
 [#5]: https://github.com/hollodotme/fast-cgi-client/issues/5
 [#6]: https://github.com/hollodotme/fast-cgi-client/issues/6
+[#9]: https://github.com/hollodotme/fast-cgi-client/issues/9